Moved RequestSimulation to dto, because it is being use as a dto, should review this later
