+++ pnnl.goss.gridappsd/src/pnnl/goss/gridappsd/data/handlers/GridLabDDataHandler.java
@@ -30,7 +30,7 @@ import pnnl.goss.core.server.DataSourceType;
 import pnnl.goss.gridappsd.api.DataManager;
 import pnnl.goss.gridappsd.api.GridAppsDataHandler;
 import pnnl.goss.gridappsd.dto.PowerSystemConfig;
-import pnnl.goss.gridappsd.requests.RequestSimulation;
+import pnnl.goss.gridappsd.dto.RequestSimulation;
 import pnnl.goss.gridappsd.utils.GridAppsDConstants;
